+++ src/main.js
@@ -1,10 +1,6 @@
 import normalizeException from 'normalize-exception'
 
-import {
-  setAggregate,
-  getAggregateErrors,
-  mergeAggregate,
-} from './aggregate.js'
+import { mergeAggregateCauses, mergeAggregateErrors } from './aggregate.js'
 import { mergeMessage } from './message.js'
 import { copyProps } from './props.js'
 import { getStackIndex, fixStack } from './stack.js'
@@ -24,39 +20,25 @@ export default function mergeErrorCause(error) {
 // own stack.
 const mergeError = function (error, stackIndex) {
   const parent = normalizeException(error)
-  const parentErrors = getAggregateErrors(parent, mergeErrorCause)
-
-  if (parent.cause === undefined) {
-    setAggregate(parent, parentErrors)
-    return parent
-  }
-
-  return mergeCause(parent, parentErrors, stackIndex)
+  mergeAggregateCauses(parent, mergeErrorCause)
+  return mergeCause(parent, stackIndex)
 }
 
 // `normalizeException()` is called again to ensure the new `name|message` is
 // reflected in `error.stack`.
-const mergeCause = function (parent, parentErrors, stackIndex) {
-  const child = mergeError(parent.cause, stackIndex - 1)
-  const message = mergeMessage(parent.message, child.message)
-  const mergedError = createError(parent, message)
-  fixStack({ mergedError, parent, child, stackIndex })
-  mergeAggregate({ mergedError, parentErrors, child, mergeErrorCause })
-  copyProps(mergedError, parent, child)
-  return normalizeException(mergedError)
-}
-
-// Ensure both the prototype and `error.name` are correct, by creating a new
-// instance with the right constructor.
-// The parent error's type is kept.
-//  - Unless its constructor throws, then we default to Error.
-const createError = function (parent, message) {
-  try {
-    return parent.constructor.name === 'AggregateError' &&
-      'AggregateError' in globalThis
-      ? new parent.constructor([], message)
-      : new parent.constructor(message)
-  } catch {
-    return new Error(message)
+const mergeCause = function (parent, stackIndex) {
+  if (parent.cause === undefined) {
+    return parent
   }
+
+  const child = mergeError(parent.cause, stackIndex - 1)
+  // eslint-disable-next-line fp/no-delete, no-param-reassign
+  delete parent.cause
+
+  mergeMessage(parent, child)
+  fixStack(parent, child, stackIndex)
+  mergeAggregateErrors(parent, child)
+  copyProps(parent, child)
+
+  return normalizeException(parent)
 }