+++ src/message.js
@@ -1,3 +1,5 @@
+import { setErrorProperty } from './set.js'
+
 // Merge parent and child error messages.
 // By default, parent messages are appended
 //  - This is because the innermost message is the most relevant one which
@@ -7,7 +9,12 @@
 // Each error message is on its own line, for clarity.
 // Empty messages are ignored
 //  - This is useful when wrapping an error properties, but not message
-export const mergeMessage = function (rawParentMessage, rawChildMessage) {
+export const mergeMessage = function (parent, child) {
+  const message = getMessage(parent.message, child.message)
+  setErrorProperty(parent, 'message', message)
+}
+
+const getMessage = function (rawParentMessage, rawChildMessage) {
   const parentMessage = rawParentMessage.trim()
   const childMessage = rawChildMessage.trim()