+++ test/aggregate.js
@@ -3,69 +3,72 @@ import mergeErrorCause from 'merge-error-cause'
 
 const { propertyIsEnumerable: isEnum } = Object.prototype
 
-if ('AggregateError' in globalThis) {
-  test('Merge cause in "errors" without "cause"', (t) => {
-    const cause = new Error('cause')
-    const innerError = new Error('innerError', { cause })
-    const error = new AggregateError([innerError], 'test')
-    const mergedError = mergeErrorCause(error)
-    t.is(mergedError.errors[0].message, 'cause\ninnerError')
-    t.false(isEnum.call(mergedError, 'errors'))
-  })
-
-  test('Merge cause in "errors" with "cause"', (t) => {
-    const cause = new Error('cause')
-    const innerError = new Error('innerError', { cause })
-    const secondCause = new Error('secondCause')
-    const error = new AggregateError([innerError], 'test', {
-      cause: secondCause,
-    })
-    const mergedError = mergeErrorCause(error)
-    t.is(mergedError.message, 'secondCause\ntest')
-    t.is(mergedError.errors[0].message, 'cause\ninnerError')
-    t.false(isEnum.call(mergedError, 'errors'))
-  })
-
-  test('Normalize "errors"', (t) => {
-    const error = new AggregateError(['innerError'], 'test')
-    t.is(mergeErrorCause(error).errors[0].message, 'innerError')
-  })
-
-  test('Handles invalid "errors"', (t) => {
-    const error = new AggregateError([], 'test')
-    error.errors = 'innerError'
-    t.is(mergeErrorCause(error).errors.length, 0)
-  })
-}
-
-test('Merge cause in "errors" without "cause" nor AggregateError', (t) => {
+test('Merge cause in "errors" without "cause"', (t) => {
   const innerError = new Error('innerError')
   // eslint-disable-next-line fp/no-mutation
   innerError.cause = new Error('cause')
   const error = new Error('test')
   error.errors = [innerError]
-  const mergedError = mergeErrorCause(error)
-  t.is(mergedError.errors[0].message, 'cause\ninnerError')
+  mergeErrorCause(error)
+  t.is(error.message, 'test')
+  t.is(error.errors[0].message, 'cause\ninnerError')
+})
+
+test('Merge cause in "errors" with "cause"', (t) => {
+  const innerError = new Error('innerError')
+  // eslint-disable-next-line fp/no-mutation
+  innerError.cause = new Error('cause')
+  const error = new Error('test')
+  error.errors = [innerError]
+  error.cause = new Error('secondCause')
+  mergeErrorCause(error)
+  t.is(error.message, 'secondCause\ntest')
+  t.is(error.errors[0].message, 'cause\ninnerError')
 })
 
 test('Does not set "errors" if none', (t) => {
   const error = new Error('test')
   error.cause = new Error('cause')
-  t.false('errors' in mergeErrorCause(error))
+  mergeErrorCause(error)
+  t.false('errors' in error)
 })
 
 test('Use child "errors" if no parent', (t) => {
   const error = new Error('test')
   error.cause = new Error('cause')
   error.cause.errors = [new Error('innerError')]
-  t.is(mergeErrorCause(error).errors[0].message, 'innerError')
+  mergeErrorCause(error)
+  t.is(error.errors[0].message, 'innerError')
+})
+
+test('Keep "errors" non-enumerable', (t) => {
+  const error = new Error('test')
+  error.cause = new Error('cause')
+  error.cause.errors = [new Error('innerError')]
+  mergeErrorCause(error)
+  t.false(isEnum.call(error, 'errors'))
+})
+
+test('Normalize "errors"', (t) => {
+  const error = new Error('test')
+  error.errors = ['innerError']
+  mergeErrorCause(error)
+  t.is(error.errors[0].message, 'innerError')
+})
+
+test('Handles invalid "errors"', (t) => {
+  const error = new Error('test')
+  error.errors = 'innerError'
+  mergeErrorCause(error)
+  t.is(error.errors, undefined)
 })
 
 test('Use parent "errors" if no child', (t) => {
   const error = new Error('test')
   error.cause = new Error('cause')
   error.errors = [new Error('innerError')]
-  t.is(mergeErrorCause(error).errors[0].message, 'innerError')
+  mergeErrorCause(error)
+  t.is(error.errors[0].message, 'innerError')
 })
 
 test('Concatenate "errors"', (t) => {
@@ -73,7 +76,8 @@ test('Concatenate "errors"', (t) => {
   error.cause = new Error('cause')
   error.errors = [new Error('one')]
   error.cause.errors = [new Error('two')]
-  t.is(mergeErrorCause(error).errors.length, 2)
-  t.is(mergeErrorCause(error).errors[0].message, 'two')
-  t.is(mergeErrorCause(error).errors[1].message, 'one')
+  mergeErrorCause(error)
+  t.is(error.errors.length, 2)
+  t.is(error.errors[0].message, 'two')
+  t.is(error.errors[1].message, 'one')
 })