+++ test/message.js
@@ -5,7 +5,8 @@ import { each } from 'test-each'
 test('Messages are trimmed', (t) => {
   const error = new Error(' test ')
   error.cause = new Error(' cause ')
-  t.is(mergeErrorCause(error).message, 'cause\ntest')
+  mergeErrorCause(error)
+  t.is(error.message, 'cause\ntest')
 })
 
 test('Empty parent messages are ignored', (t) => {
@@ -13,9 +14,9 @@ test('Empty parent messages are ignored', (t) => {
   const error = new Error('')
   error.prop = true
   error.cause = new Error('cause')
-  const { message, prop } = mergeErrorCause(error)
-  t.is(message, 'cause')
-  t.true(prop)
+  mergeErrorCause(error)
+  t.is(error.message, 'cause')
+  t.true(error.prop)
 })
 
 test('Empty child messages are ignored', (t) => {
@@ -23,33 +24,37 @@ test('Empty child messages are ignored', (t) => {
   // eslint-disable-next-line unicorn/error-message
   error.cause = new Error('')
   error.cause.prop = true
-  const { message, prop } = mergeErrorCause(error)
-  t.is(message, 'test')
-  t.true(prop)
+  mergeErrorCause(error)
+  t.is(error.message, 'test')
+  t.true(error.prop)
 })
 
 each(['test:', 'test: '], ({ title }, message) => {
   test(`Prepend messages with colon | ${title}`, (t) => {
     const error = new Error(message)
     error.cause = new Error('cause')
-    t.is(mergeErrorCause(error).message, 'test: cause')
+    mergeErrorCause(error)
+    t.is(error.message, 'test: cause')
   })
 })
 
 test('Prepend messages with colon and newline', (t) => {
   const error = new Error('test:\n')
   error.cause = new Error('cause')
-  t.is(mergeErrorCause(error).message, 'test:\ncause')
+  mergeErrorCause(error)
+  t.is(error.message, 'test:\ncause')
 })
 
 test('Does not prepend messages with newline but no colon', (t) => {
   const error = new Error('test\n')
   error.cause = new Error('cause')
-  t.is(mergeErrorCause(error).message, 'cause\ntest')
+  mergeErrorCause(error)
+  t.is(error.message, 'cause\ntest')
 })
 
 test('New error message is reflected in stack', (t) => {
   const error = new Error('test')
   error.cause = new Error('cause')
-  t.true(mergeErrorCause(error).stack.includes('cause\ntest'))
+  mergeErrorCause(error)
+  t.true(error.stack.includes('cause\ntest'))
 })